ci: run Jest tests in the CI workflow
Run Jest tests as part of the CI workflow. Because the AWS Cloud
Development Kit allows us to write our CloudFormation stack definitions
in TypeScript, we can write Jest tests against the code. See
https://amzn.to/2HnBSaw for more details.

The job configuration itself is inspired by the work done in
ShahradR/action-taskcat. For details on the workflow changes, see
commits in that repository.